Still refactoring the YAP Prover code. Found a couple of bugs in the way the queries to YAP Prolog are managed: some methods don't manage the case when the query fails and some methods don't manage the case when there is expected to be only one result for the query, but due to unforeseen causes there are more. Moreover, when initialization fails, the fact that the current prolog state on the Java side has not been initialized creates NullPointerException in consecutive calls to the state machine.
Added StateMachineException to be thrown when a method of the state machine fails for various reasons (e.g. query taking to long to be computed, state of the state machine is inconsistent,...)

Added GGPBaseProverVerifier.java to check how many iterations of the state machine tests can be made when comparing two ProverStateMachine.java state machines.

Added the TODO file with all fixes to be made to the project.